Refactor host code
Remove postform.rs and move it to lib.rs. The previous functions in
lib.rs should actually belong to the application in order to decouple
the application from the library.
