+++ postform-host/src/main.rs
@@ -1,14 +1,21 @@
 use cobs::CobsDecoder;
-use postform_host::{download_firmware, run_core, ElfMetadata, LogLevel};
-use probe_rs::config::registry;
-use probe_rs::Probe;
+use color_eyre::eyre::Result;
+use object::read::{File as ElfFile, Object, ObjectSymbol};
+use postform_host::{ElfMetadata, LogLevel};
+use probe_rs::{
+    config::registry,
+    flashing::{download_file, FileDownloadError, Format},
+    Probe, Session,
+};
 use probe_rs_rtt::Rtt;
-use std::path::PathBuf;
-use std::sync::{Arc, Mutex};
+use std::{
+    fs,
+    path::PathBuf,
+    sync::{Arc, Mutex},
+    time::Duration,
+};
 use structopt::StructOpt;
 use termion::color;
-
-use color_eyre::eyre::Result;
 
 fn print_probes() {
     let probes = Probe::list_all();
@@ -36,7 +43,7 @@ fn print_chips() {
 }
 
 #[derive(Debug, StructOpt)]
-#[structopt(name = "deferred-logger")]
+#[structopt()]
 struct Opts {
     /// List supported chips and exit.
     #[structopt(long)]
@@ -53,6 +60,48 @@ struct Opts {
     /// Path to an ELF firmware file.
     #[structopt(name = "ELF", parse(from_os_str), required_unless_one(&["list-chips", "list-probes"]))]
     elf: Option<PathBuf>,
+}
+
+pub fn download_firmware(session: &Arc<Mutex<Session>>, elf_path: &PathBuf) {
+    let mut mutex_guard = session.lock().unwrap();
+    println!("Loading FW to target");
+    if let Err(error) = download_file(&mut mutex_guard, &elf_path, Format::Elf) {
+        match error {
+            FileDownloadError::Elf(_) => {
+                println!("Error with elf file");
+            }
+            FileDownloadError::Flash(_) => {
+                println!("Error flashing the device");
+            }
+            _ => {
+                println!("Other error downloading FW");
+            }
+        }
+        return;
+    }
+
+    let file_contents = fs::read(elf_path).unwrap();
+    let elf_file = ElfFile::parse(&file_contents[..]).unwrap();
+    let main = elf_file
+        .symbols()
+        .find(|s| s.name().unwrap() == "main")
+        .expect("main symbol not found!");
+
+    let mut core = mutex_guard.core(0).unwrap();
+    let _ = core.reset_and_halt(Duration::from_millis(10)).unwrap();
+    core.set_hw_breakpoint(main.address() as u32)
+        .expect("Unable to set breakpoint in main");
+    core.run().unwrap();
+    core.wait_for_core_halted(Duration::from_secs(1))
+        .expect("Didn't halt on main");
+    println!("Download complete!");
+}
+
+pub fn run_core(session: Arc<Mutex<Session>>) {
+    let mut mutex_guard = session.lock().unwrap();
+    let mut core = mutex_guard.core(0).unwrap();
+    core.clear_all_hw_breakpoints().unwrap();
+    core.run().unwrap();
 }
 
 fn color_for_level(level: LogLevel) -> String {
